Moved loading of indoor locations into a singleton to avoid long load times. Improved generation of place ID for indoor location.
