+++ app/src/main/java/com/example/campusguide/search/indoor/Building.kt
@@ -1,0 +1,17 @@
+package com.example.campusguide.search.indoor
+
+import com.beust.klaxon.Json
+
+data class Building(
+    @Json(name = "name")
+    val name: String,
+
+    @Json(name = "code")
+    val code: String,
+
+    @Json(name = "address")
+    val address: String,
+
+    @Json(name = "rooms")
+    val rooms: List<Room>
+)
